refactor: changes in names
change news by publications
